Hackathon/requirements.txt
altgraph==0.17.2
blis==0.7.5
branca==0.4.2
catalogue==2.0.6
certifi==2021.10.8
charset-normalizer==2.0.7
click==8.0.3
cluster==1.4.1.post3
colorama==0.4.4
cycler==0.11.0
cymem==2.0.6
DAWG-Python==0.7.2
docopt==0.6.2
en-core-web-sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.1.0/en_core_web_sm-3.1.0-py3-none-any.whl
et-xmlfile==1.1.0
folium==0.12.1
future==0.18.2
geographiclib==1.52
geopy==2.2.0
idna==3.3
Jinja2==3.0.2
kiwisolver==1.3.2
MarkupSafe==2.0.1
matplotlib==3.4.3
murmurhash==1.0.6
numpy==1.21.3
openpyxl==3.0.9
packaging==21.2
pandas==1.3.4
pathy==0.6.1
pefile==2021.9.3
Pillow==8.4.0
preshed==3.0.6
pydantic==1.8.2
pyinstaller==4.6
pyinstaller-hooks-contrib==2021.3
pymorphy2==0.9.1
pymorphy2-dicts-ru==2.4.417127.4579844
pyparsing==2.4.7
PyQt5==5.15.6
PyQt5-Qt5==5.15.2
PyQt5-sip==12.9.0
PyQt5-stubs==5.15.2.0
PyQtWebEngine==5.15.5
PyQtWebEngine-Qt5==5.15.2
python-dateutil==2.8.2
pytz==2021.3
pywin32-ctypes==0.2.0
requests==2.26.0
scipy==1.7.1
seaborn==0.11.2
six==1.16.0
smart-open==5.2.1
spacy==3.1.4
spacy-legacy==3.0.8
srsly==2.4.2
thinc==8.0.12
tornado==6.1
tqdm==4.62.3
typer==0.4.0
typing-extensions==3.10.0.2
urllib3==1.26.7
wasabi==0.8.2
